Organized code and bug fixes
- Refactored code so it's more manageable.
- Fixed item list request being sent twice.

TODO:
*implement notifications
*implement payment gateway
*implement orders archive
*implement drawerlayout + view account
