+++ app/src/main/AndroidManifest.xml
@@ -4,12 +4,12 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <application
         android:allowBackup="true"
-        android:name=".MyApplication"
+        android:name=".utils.MyApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -17,13 +17,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ItemDetailActivity"
-            android:parentActivityName=".MainActivity" />
-        <activity android:name=".CartActivity"
-            android:parentActivityName=".MainActivity"/>
-        <activity android:name=".Login" android:theme="@style/AppTheme.Dark"
+            android:name=".activities.ItemDetailActivity"
+            android:parentActivityName=".activities.MainActivity" />
+        <activity android:name=".activities.CartActivity"
+            android:parentActivityName=".activities.MainActivity"/>
+        <activity android:name=".activities.Login" android:theme="@style/AppTheme.Dark"
             android:windowSoftInputMode="stateAlwaysHidden"/>
-        <activity android:name=".Register" android:theme="@style/AppTheme.Dark" />
+        <activity android:name=".activities.Register" android:theme="@style/AppTheme.Dark" />
     </application>
 
 </manifest>
