+++ app/src/main/java/com/example/acfan/project/fragments/ItemDetailFragment.java
@@ -1,4 +1,4 @@
-package com.example.acfan.project;
+package com.example.acfan.project.fragments;
 
 import android.app.Fragment;
 import android.content.Intent;
@@ -12,16 +12,22 @@ import android.widget.TextView;
 
 import com.android.volley.toolbox.ImageLoader;
 import com.android.volley.toolbox.NetworkImageView;
+import com.example.acfan.project.R;
+import com.example.acfan.project.network.VolleySingleton;
+import com.example.acfan.project.activities.CartActivity;
+import com.example.acfan.project.models.Cart;
+import com.example.acfan.project.models.Item;
+import com.example.acfan.project.utils.CartHelper;
 
 /**
  * Created by acfan on 10/22/2016.
  */
 
 public class ItemDetailFragment extends Fragment {
-    public static final String ITEM_CART ="com.example.acfan.project.Item Cart";
+    public static final String ITEM_CART ="com.example.acfan.project.models.Item Cart";
     private static final String TAG ="ItemDetailFragment";
     private int count=1;
-    public static final String ITEM_AMOUNT_EXTRA="com.example.acfan.project.Item Amount";
+    public static final String ITEM_AMOUNT_EXTRA="com.example.acfan.project.models.Item Amount";
     Intent intent;
     Item item;
     TextView name,description,price,amount;
@@ -96,6 +102,7 @@ public class ItemDetailFragment extends Fragment {
                 cart.addItem(count,item);
                 Intent intent = new Intent(getActivity(), CartActivity.class);
                 startActivity(intent);
+                getActivity().finish();
             }
         });
     }
