+++ app/src/main/java/com/example/acfan/project/models/Cart.java
@@ -1,4 +1,4 @@
-package com.example.acfan.project;
+package com.example.acfan.project.models;
 
 import android.os.Parcel;
 import android.os.Parcelable;
