+++ app/src/main/java/com/example/acfan/project/network/CustomJsonArrayRequest.java
@@ -1,4 +1,4 @@
-package com.example.acfan.project;
+package com.example.acfan.project.network;
 
 import com.android.volley.NetworkResponse;
 import com.android.volley.ParseError;
