+++ app/src/main/java/com/example/acfan/project/utils/CartHelper.java
@@ -1,4 +1,6 @@
-package com.example.acfan.project;
+package com.example.acfan.project.utils;
+
+import com.example.acfan.project.models.Cart;
 
 /**
  * A helper class to retrieve the static shopping cart. Call {@code getCart()} to retrieve the shopping cart before you perform any operation on the shopping cart.
