+++ app/src/main/res/layout/activity_main.xml
@@ -3,10 +3,11 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
+    android:id="@+id/constraint_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.acfan.project.MainActivity">
+    tools:context="com.example.acfan.project.activities.MainActivity">
+    <!--
     <include
         layout="@layout/home_page"
         android:layout_height="0dp"
@@ -18,5 +19,30 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintVertical_bias="1.0" />
+        !-->
+    <FrameLayout
+        android:layout_width="368dp"
+        android:layout_height="410dp"
+        android:id="@+id/list"
+        tools:layout_editor_absoluteY="8dp"
+        tools:layout_editor_absoluteX="8dp">
+
+    </FrameLayout>
+    <android.support.design.widget.FloatingActionButton
+        android:onClick="OpenCart"
+        app:backgroundTint="#e0e0e0"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:clickable="true"
+        app:fabSize="normal"
+        app:srcCompat="@drawable/cart_fab"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentEnd="true"
+        android:layout_marginEnd="15dp"
+        android:id="@+id/floatingActionButton"
+        android:layout_marginRight="15dp"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginBottom="29dp" />
 
 </android.support.constraint.ConstraintLayout>
