--- app/src/main/res/layout/home_page.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    <fragment
-        android:id="@+id/item_frag"
-        android:name="com.example.acfan.project.Item_Frag"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-    </fragment>
-
-    <android.support.design.widget.FloatingActionButton
-        android:onClick="OpenCart"
-        app:backgroundTint="#e0e0e0"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:clickable="true"
-        app:fabSize="normal"
-        app:srcCompat="@drawable/cart_fab"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginEnd="15dp"
-        android:layout_marginBottom="25dp"
-        android:id="@+id/floatingActionButton" />
-
-</RelativeLayout>
